add xml comments to input field andreorganize code
